backend/requirements.txt
fastapi
uvicorn
pydicom
numpy
pillow
python-multipart